runtime: integrate tracer into inport and outport constructors for improved message tracking
